+++ weaver/sort.py
@@ -7,6 +7,14 @@ SORT_USER = "user"
 SORT_QUOTE = "quote"
 SORT_PRICE = "price"
 SORT_ID = "id"
+SORT_ID_LONG = "identifier"  # long form employed by Processes in DB representation
+
+PROCESS_SORT_VALUES = frozenset([
+    SORT_ID,
+    SORT_ID_LONG,  # will replace by short ID to conform with JSON representation
+    SORT_PROCESS,  # since listing processes, can be an alias to ID
+    SORT_CREATED,
+])
 
 JOB_SORT_VALUES = frozenset([
     SORT_CREATED,